使course页面接入api，并精简组件
